fix: url parsing issue with wishlists
Refers to #27 - we now accept the new format for wishlists as well. But since Geizhals uses serverside rendering or some clientside javascript magic, we'd need to write some code to access the api directly. This takes quite a bit to implement so for now we need to live without the wishlists.